+++ src/shared/types/api/common.ts
@@ -9,3 +9,7 @@ export interface GetUrlNameRes {
 		tabName: string;
 	};
 }
+
+export interface PostToggleTaskStatusReq {
+	taskId: number;
+}
